Reintroduce RootViewModel for state updates and session handling
